backend/pyproject.toml
[tool.poetry]
name = "dih-champions-backend"
version = "0.1.0"
description = "Django backend for DIH Champions"
authors = ["DIH Champions Team"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
django = ">=5.0,<6.0"
djangorestframework = ">=3.15,<4.0"
django-cors-headers = ">=4.4,<5.0"

[tool.poetry.group.dev.dependencies]

[build-system]
requires = ["poetry-core>=1.9.0"]
build-backend = "poetry.core.masonry.api"
